internal/metrics moved into pkg/metrics
